move stack to dto
