+++ lib/DTO/Stack.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Einenlum\PhpStackDetector;
+namespace Einenlum\PhpStackDetector\DTO;
+
+use Einenlum\PhpStackDetector\StackType;
 
 readonly class Stack
 {
