+++ lib/Detector.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Einenlum\PhpStackDetector;
 
+use Einenlum\PhpStackDetector\DTO\Stack;
+
 readonly class Detector
 {
     /** @param StackDetectorInterface[] $stackDetectors */
