+++ lib/StackDetectorInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Einenlum\PhpStackDetector;
 
+use Einenlum\PhpStackDetector\DTO\Stack;
+
 interface StackDetectorInterface
 {
     public function getStack(string $baseUri, ?string $subDirectory): ?Stack;
